Sometimes I want to install ohmyzsh from github

+++ build_development_environment.sh
@@ -56,17 +56,29 @@ contains $* "--upgrade-py" && echo "---- Optional: install python3.11" && ./upgr
 
 echo "---- Optional: install zsh"
 sudo apt install -y zsh
-sudo python -c "fn='/etc/pam.d/chsh';a,b='required','sufficient';fp=open(fn);s=fp.read();fp.close();fp=open(fn,'w');fp.write(s.replace(a,b));fp.close()"
-# sudo python -c "fn='/etc/pam.d/chsh';a,b='required','sufficient';fp=open(fn,'a+');s=fp.read();fp.truncate();fp.write(s.replace(b,a));fp.close()"
+if [ -f /etc/pam.d/chsh ]; then
+  # Make `chsh` no need to input password
+  # sudo python -c "fn='/etc/pam.d/chsh';a,b='required','sufficient';fp=open(fn,'a+');s=fp.read();fp.truncate();fp.write(s.replace(b,a));fp.close()"
+  sudo python -c "fn='/etc/pam.d/chsh';a,b='required','sufficient';fp=open(fn);s=fp.read();fp.close();fp=open(fn,'w');fp.write(s.replace(a,b));fp.close()"
+fi
 chsh -s $(which zsh)
 sh -c 'echo "[ -s \$HOME/.bash_aliases ] && source \$HOME/.bash_aliases" >> $HOME/.zshrc'
 sh -c 'echo "[ -s \$HOME/.local/bin ] && export PATH=\$HOME/.local/bin:/usr/local/bin:\$PATH" >> $HOME/.zshrc'
 sh -c 'echo "export ZSH=\$HOME/.oh-my-zsh" >>  $HOME/.zshrc'
 sh -c 'echo "ZSH_THEME=random" >>  $HOME/.zshrc'
-sh -c 'echo "plugins=(git pip python pipenv)" >>  $HOME/.zshrc'
+sh -c 'echo "plugins=(git pip python poetry)" >>  $HOME/.zshrc'
 sh -c 'echo "[ -s \$ZSH/oh-my-zsh.sh ] && source \$ZSH/oh-my-zsh.sh" >>  $HOME/.zshrc'
 
-export REMOTE="https://gitee.com/mirrors/oh-my-zsh.git" && sh -c "$(curl -fsSL https://gitee.com/mirrors/oh-my-zsh/raw/master/tools/install.sh)" --keep-zshrc
+if [ -v OMZ_REPO ]; then
+  echo use $OMZ_REPO for oh-my-zsh installing
+else
+  if [ -v USE_GITHUB ]; then
+    OMZ_REPO="https://github.com/ohmyzsh/oh-my-zsh"
+  else
+    OMZ_REPO="https://gitee.com/mirrors/oh-my-zsh"
+  fi
+fi
+export REMOTE="$OMZ_REPO.git" && sh -c "$(curl -fsSL $OMZ_REPO/raw/master/tools/install.sh)" --keep-zshrc
 
 
 echo "---- Init python development environment."
